GalleryView refactor
